THAER_CV/TDK_website/requirements.txt
accelerate==0.30.0
beautifulsoup4==4.12.3
blinker==1.8.1
bottle==0.12.25
bs4==0.0.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diffusers==0.27.2
filelock==3.14.0
Flask==3.0.3
fsspec==2024.3.1
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.1.0
inflect==7.2.1
intel-openmp==2021.4.0
iso639-lang==2.2.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
package-name==0.1
packaging==24.0
pillow==10.3.0
psutil==5.9.8
PyMultiDictionary==1.2.4
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.3
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sympy==1.12
tbb==2021.12.0
torch==2.3.0
tqdm==4.66.4
typeguard==4.2.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.1
